// src/app/Shared/models/life/setup/organization/appraisal/subdivision.ts
export class Subdivision {
    "appraisalCode": string;
    "auth": number;
    "authBy": string;
    "authOn": Date;
    "companyCode": string;
    "createdBy": string;
    "createdOn": Date;
    "criteria1": string;
    "criteria1Cond": string;
    "criteria1ProdClass": string;
    "criteria1Score": number;
    "criteria1Target": number;
    "criteria1Weight": number;
    "criteria2": string;
    "criteria2Cond": string;
    "criteria2ProdClass": string;
    "criteria2Score": number;
    "criteria2Target": number;
    "criteria2Weight": number;
    "criteria3": string;
    "criteria3Cond": string;
    "criteria3ProdClass": string;
    "criteria3Score": number;
    "criteria3Target": number;
    "criteria3Weight": number;
    "criteria4": string;
    "criteria4Cond": string;
    "criteria4ProdClass": string;
    "criteria4Score": number;
    "criteria4Target": number;
    "criteria4Weight": number;
    "description": string;
    "divPerfWeight": number;
    "id": number;
    "subDivPerfWeight": number;
    "versionDate": Date;
    "versionNo": number
  }